KMI Whitelists: allow alternative section suffix 'symbol_list'
As of now, KMI whitelists are defined as ini-file style sections named
with trailing "whitelist". A number of projects have recently introduced
policies that discourage the use of terminology with negative
connotations in an effort to make the language used in their code base
more inclusive. One of such terms is 'whitelist'. In order to support
projects choosing to avoid this term, add support for an alternative:
"symbol_list". This allows e.g. a section name "[abi_symbol_list]".

While commonly chosen alternatives are "allow_list", "include_list" or
similar, I intentionally took out any positive/negative meaning as I
think those are anyway misleading. (Is the whitelist referring to symbols
that are allowed to change or is it referring to the list of symbols
that are kept stable?) In an effort to make this more obvious, I went for
"symbol list" to describe the symbols that are part of the ABI analysis
either during extraction or during comparison.

This is a backwards compatible change as we still support the prior
suffix. I adjusted documentation and added tests accordingly.

	* doc/manuals/kmidiff.rst: Adjust documentation.
	* src/abg-tools-utils.cc (gen_suppr_spec_from_kernel_abi_whitelists):
	  add support for additional whitelist section suffix.
	* tests/data/Makefile.am: Add new test files.
	* tests/data/test-kmi-whitelist/symbol-list-with-another-single-entry: New test case.
	* tests/data/test-kmi-whitelist/symbol-list-with-duplicate-entry: Likewise.
	* tests/data/test-kmi-whitelist/symbol-list-with-single-entry: Likewise.
	* tests/data/test-kmi-whitelist/symbol-list-with-two-sections: Likewise.
	* tests/test-kmi-whitelist.cc: Wire up new test cases.

Reviewed-by: Giuliano Procida <gprocida@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ied5a16e4c16abe622622db0aeee65a42032d9e48