Drop unnecessary upstream files that impose RESTRICTED licensing
ltmain.sh and gen-changelog.py are maintainer scripts that are not
required for downstream projects such as Android. Their licensing would
otherwise taint the project to be restricted. We can safely drop them as
we do not need them and since they hardly ever change upstream, we also
do not run into merge conflicts.

Bug: 151953481
Bug: 151177513
Bug: 67772237
Change-Id: I25b2d9bf98b930a4d7bfe60bda9b35237c468241
Signed-off-by: Matthias Maennich <maennich@google.com>