abitidy: drop empty elements
Certain elements in ABI XML are effectvely containers whose attributes
do not carry ABI information. They can be dropped if empty.

- elf-variable-symbols: pure container
- elf-function-symbols: pure container
- namespace-decl: C++ namespace (has a name attribute)
- abi-instr: compilation unit (has path etc.)
- abi-corpus: binary object (has architecture)
- abi-corpus-group: binary objects (has architecture)

It could be argued that abi-corpus (or abi-corpus-group) should be
kept around to hold the architecture of an object or set of objects.
However, if a binary object has no symbols (say, if it is empty), it
hardly matters what the architecture is.

Empty elements can also be produced during manipulation of the XML.

Note that abidiff currently rejects completely empty files so we avoid
dropping empty top-level elements.

	* tools/abitidy.cc (from_libxml): Cast a libxml string to C
	string.
	(remove_element): Remove an XML element and any immediately
	preceding comment.
	(DROP_IF_EMPTY): Set of names of elements that are safe to
	drop if empty.
	(drop_empty): Drop empty elements, if safe to do so,
	recursively.
	(main): Add drop-empty option processing and action.

Bug: 181269148
Change-Id: Ie728522026210d48a49da99bbd453c0be8139721
Signed-off-by: Giuliano Procida <gprocida@google.com>