abitidy: prune unreachable elements
This commit adds a pass to prune unreachable parts of the ABI.

This has been done primarily to tackle the issue of very large ABIs
for libraires implemented in C++ but exposing only a C API.

The aim is to remove parts of the XML that are irrelevant to
abidiff. ELF symbols (and shared object dependencies) are taken to be
the graph roots for the purpose of reachability analysis.

On most ABI XML files, this results in a modest saving (of 5-10%).
However, for a pathological example the resulting XML is more than
2300 times smaller.

The implementation is as follows:

- the entire XML document is traversed and a digraph is created
  - vertices are either types or symbol declarations
  - edges represent contains, containing and refers-to relationships
- starting from the ELF symbols, all reachable vertices are recorded
- unreachable types and symbol declarations are dropped

	* tools/abitidy.cc (get_attribute): Fetch an attribute.
	(prune_unreachable): Prune unreachable elements.
	(main): Add prune-unreachable option handling and action.

Bug: 181269148
Change-Id: Ice327606f9b9ce03ed179fb0fe8d03989828d066
Signed-off-by: Giuliano Procida <gprocida@google.com>