abitidy: sort namespaces, types and declarations
ABI XML as generated by abidw is sensitive to processing order and in
particular the placement of types and declarations in the XML is
unpredictable. This leads to difficulties in keeping ABI XML under
version control.

This commit adds a pass which consolidates all namespaces, types and
declarations within an abi-corpus into a replacement abi-instr. These
elements are ordered by name, id and name respectively, with types
before declarations and declarations before nested namespaces.

Note that any annotation comments for namespaces are lost as a
side-effect of the namespace element consolidation.

If duplicate type definitions spanning multiple namespace scopes are
present, the sorting pass is skipped to avoid the risk of changing the
interpretation of such malformed ABI XML.

All other duplicate definitions types and declarations have their
relative order preserved - the sort is stable.

A notable example of a type with conflicting definitions is type-id-86
in tests/data/test-read-dwarf/PR22122-libftdc.so.abi in the libabigail
test suite.

	* tools/abitidy.cc (move_node): Move a node to an
	element.
	(move_element): Move an XML element and any immediately
	preceding comment to another element.
	(INSTR_VARIABLE_ATTRIBUTES): Set of abi-instr attributes which
	are expected to vary within an abi-corpus.
	(get_children_by_namespace): Collect elements of an abi-instr
	by namespace scope.
	(sort_namespaces_types_and_declarations): Sort namespaces,
	types and declarations within an abi-corpus.
	(main): Add sort option processing and action. Check for
	conflicting type definitions if sorting requested.

Bug: 181269148
Change-Id: Iaf75c2f3a46b2c59ced94a35faefd3f3d77ca17f
Signed-off-by: Giuliano Procida <gprocida@google.com>