abg-reader: Ensure corpus always has a symtab reader
In the presence of an empty abi-corpus element and with the following
change to always allocate a fresh corpus object, such objects can
sometimes be left without a symtab reader, instead of inheritng one
from the previous corpus.

The reader is called to obtain sorted lists of symbols during ABI
comparisons. The simplest way to avoid a crash is to maintain the
invariant that a reader object is always present.

With this change, if there is bad XML preventing symbols from being
read, no error is raised as before, but the logic has been tweaked so
that abi-instr parsing will nevertheless be attempted.

	* src/abg-reader.cc (read_symbol_db_from_input): Fix
	documentation for this function. Allow "successful parsing" to
	include the case where no symbols were present in the input.
	(read_corpus_from_input): Unconditionally set a symtab reader
	on the corpus object. Unconditionally parse the abi-instr of a
	corpus.

Signed-off-by: Giuliano Procida <gprocida@google.com>
(cherry picked from commit 25bd77e31ea688c91669db9e9dd8d80f0ba0af62)
Bug: 187831743
Change-Id: If448934c397dae1869092a56a757e131b6439f57
Signed-off-by: Giuliano Procida <gprocida@google.com>