abitidy: reanonymise named anonymous types
The attribute naming-typedef-id is a backwards link from an anonymous
type to the typedef that refers to it. It is ignored by abidiff.

A recent change to abidw changed its behaviour for any anonymous type
that has a naming typedef. In addition to linking the typedef and type
in both directions, the code now gives (some) anonymous types the same
name as the typedef. This misrepresents the original type information.

With the option --reanonymise-anonymous or -r, also implied by --all,
such types are reanonymised.

This change also merges the handling of anonymous types into a single
traversal of the type XML tree. This is more efficient than 3 separate
traversals.

Additionally, it adjusts the detection of anonymous types to use the
is-anonymous attribute. This is safer than pattern matching on the
name as libabigail has changed this in past.

Finally, it now even handles the case where the name may be missing
altogether.

Old abidw output:

```
    <typedef-decl name='kprojid_t' type-id='393e7ae8' id='b7cdd2f8'/>
    <class-decl name='__anonymous_struct__63' size-in-bits='32' is-struct='yes' is-anonymous='yes' naming-typedef-id='b7cdd2f8' visibility='default' id='393e7ae8'>
      <data-member access='public' layout-offset-in-bits='0'>
        <var-decl name='val' type-id='6ab42ed1' visibility='default'/>
      </data-member>
    </class-decl>
```

New abidw output:

```
    <typedef-decl name='projid_t' type-id='70734f24' id='6ab42ed1'/>
    <class-decl name='kprojid_t' size-in-bits='32' is-struct='yes' naming-typedef-id='b7cdd2f8' visibility='default' id='02ad2737'>
      <data-member access='public' layout-offset-in-bits='0'>
        <var-decl name='val' type-id='6ab42ed1' visibility='default'/>
      </data-member>
    </class-decl>
```

Previously, abitidy --all would have removed the 63 suffix and the
naming-typedef-id. It would have left the second struct with an
incorrect name and no is-anonymous flag.

Now it produces the following (modulo ordering), respectively:

```
    <typedef-decl name='kprojid_t' type-id='393e7ae8' id='b7cdd2f8'/>
    <class-decl name='__anonymous_struct__' size-in-bits='32' is-struct='yes' is-anonymous='yes' visibility='default' id='393e7ae8'>
      <data-member access='public' layout-offset-in-bits='0'>
        <var-decl name='val' type-id='6ab42ed1' visibility='default'/>
      </data-member>
    </class-decl>
```

and

```
    <typedef-decl name='kprojid_t' type-id='02ad2737' id='b7cdd2f8'/>
    <class-decl name='__anonymous_struct__' size-in-bits='32' is-struct='yes' visibility='default' id='02ad2737' is-anonymous='yes'>
      <data-member access='public' layout-offset-in-bits='0'>
        <var-decl name='val' type-id='6ab42ed1' visibility='default'/>
      </data-member>
    </class-decl>
```

	* tools/abitidy.cc (discard_naming_typedefs): Fold this
	functionality into new function handle_anonymous_types.
	(normalise_anonymous_type_names): Likewise.
	(handle_anonymous_types): New function with the same structure
	as the old normalise_anonymous_type_names. Recognise
	explicitly anonymous types by their is-anonymous attribute.
	Types that may need to be reanonymised are those with
	naming-typedef-id attributes. When normalising or
	reanonymising, set the name to the expected name. When
	reanonymising, set the is-anonymous attribute. Discarding
	naming-typedef-id attributes is much as before, though element
	selection is now shared with the other two operations.

Bug: 193031982
Change-Id: I5cd770a006d099a0e3cd92bb4f51dfa1c1a56817
Signed-off-by: Giuliano Procida <gprocida@google.com>