abitidy: split out unset_attribute functionality
At some point in the past it seemed it might be useful to have a
single function look up and set or unset an attribute value and use an
optional value for uniformity.

However, none of the callers do anything other than construct a new
optional which is then immediately unpacked (if not empty) by the
function.

This change splits set_attribute and unset_attribute into separate
functions and adjusts all callers.

Change-Id: I8134c8f2455a9232f2aae326b8d8e0d4d8252b62
Signed-off-by: Giuliano Procida <gprocida@google.com>