abitidy: fix up symbol references after symbol removal
ABI XML models symbol <-> symbol and symbol <-> declaration
relationships in a way that makes symbol removal a non-trivial
operation in the presence of aliasing.

If a group of symbols are aliased, libabigail picks one as the main
symbol and refers to the others as aliases.

Note that things get even more complicated with libabigail symbol
suppression. However, this change will allow us to stop using that
functionality.

In the XML:

* a main symbol has a comma-separated list of aliases
* any declaration is associated with the main symbol

This change ensures that:

* a new main symbol is picked if a main symbol is removed and at least
one alias remains
* alias lists are rewritten / removed
* references in declarations are rewritten / removed

To ensure deterministic behaviour, updated alias lists are kept sorted
and the first available alias is promoted to be the new main symbol.

Bug: 232406384
Change-Id: Iddf02e1445866b2e1ed807b50bda65a53131f9ef
Signed-off-by: Giuliano Procida <gprocida@google.com>