abitidy: fix lifetime issues in filter_symbols
The previous change introduced a potential iterator invalidation issue
and a definite use-after-free issue. Both were caused by holding
references to map values. Non-optimised builds crashed.

This change eliminates the first issue by swapping the relevant value
out of and into the map. The use-after-free is fixed by ensuring the
key is deleted after the last use of the value.

Bug: 232406384
Change-Id: I3f8997d685485219b5a8a74f6d38545833ae1cd7
Signed-off-by: Giuliano Procida <gprocida@google.com>