abitidy: Renumber anonymous type ids using option `renumber-anonymous-types`
Anonymous type renumbering pass can be done before or after the existing
anonymous type handling pass, but doing it beforehand ensures that it
will be applied to a smaller set of types. This is because types with
naming typedef backlinks will be excluded and libabigail has in any case
likely produced unique ids for these.

This option is experimental for now and not part of the `--all` options.
When the flag is used, type ids for anonymous types are renumbered by
aditidy using information about the type making it less to random
changes due to linear probing during type id generation.

Bug: 239910948
Signed-off-by: Sid Nayyar <sidnayyar@google.com>
Change-Id: Ibbfdd4bb6eec14d91e85942b7580e5427981e692