Bug 30461 - insight fails self-compare
This self-comparison issue is due to several different underlying
problems.

Class destructors are often implemented by the compiler using cloned
functions, even for virtual destructors.  In DWARF, for a given class,
the cloned function implementing the virtual destructor might come way
after the point where the class and its member functions (including
the destructors) have been declared.  So the DWARF reader needs to
match the "dangling" cloned function that implements the virtual
destructor with the destructor declared in the class declaration.
Note that it's the cloned definition that contains the address of the
ELF symbol for the destructor, not the declaration.  So we really need
to see the cloned function to build a proper IR (that also represents
the ELF symbol) for the destructor.

As the destructor is implemented by a cloned function, we might see
(in the DWARF) other clones of that function that have the same
linkage name.  In that case, if the IR of the class already has a
destructor with the same linkage name, then we not build another IR
for it, otherwise we'd wrongly duplicate the destructor.  And this is
exactly the error we were doing in this issue.  This patch addresses
that issue.

The ABIXML does unfortunately have classes with duplicated
destructors, due to this.  So as the DWARF reader is now going to be
fixed, the ABIXML reader also needs to be fixed to avoid reading
duplicated member functions.  The patch addresses that issue as well.

Then, when the above was fixed, I stumbled across an issue related to
type fundamentals of type canonicalization:

I think we will ultimately need to be able to canonicalize types in
the same order, if they come from DWARF or ABIXML.  It now appears to
me that because of recursive types in the context of ODR-violations
(two different types being present in different translation units and
yet having the same name.  For instance, the BFD type from binutils
and this insight package), type canonicalization is not a commutative
operation, unfortunately.  That is why we need to either store some
type hash in ABIXML to avoid having to re-do type canonicalization
when reading ABIXML, or canonicalize types coming from ABIXML in the
same order as they were, when they came from their initial format
(DWARF or otherwise).  This would be a project in it own right.  Until
then, a workaround that seems to be enough in this case is to clear
the type comparison result cache after canonicalizing each type.

The rest of the patch adjusts the regression tests output as needed.

For the record, below is the command line that triggered the issue:

    $ fedabipkgdiff --debug --self-compare -a --from fc38 insight

	* src/abg-dwarf-reader.cc (build_ir_node_from_die): If we are
	looking at a cloned function that's a member function to be added
	to an existing class, make sure a member function with the same
	linkage name doesn't already exist in the class before adding this
	one.  Otherwise, we'd be duplicating a member function inside the
	class.
	* src/abg-ir.cc (method_decl::set_linkage_name): When setting the
	linkage name of a method to a new one, erase the old method that
	had the old linkage name from the containing class.
	(compare_canonical_type_against_candidate): Clear the comparison
	type result cache after each type canonicalization to avoid
	re-using cached result that should have been invalidated.  This is
	a work-around the more fundamental type canonicalization issue
	outlined in the preamble of this commit log.
	* src/abg-reader.cc (build_function_decl): Avoid loading
	duplicated member functions.  The key of the function being its
	mangled name.
	(build_class_decl): The XML node to map when looking a member
	function is really the XML node for the member function.
	* tests/data/test-abidiff/test-PR18791-report0.txt: Adjust.
	* tests/data/test-read-dwarf/PR22015-libboost_iostreams.so.abi:
	Likewise.
	* tests/data/test-read-dwarf/PR22122-libftdc.so.abi: Likewise.
	* tests/data/test-read-dwarf/test-libandroid.so.abi: Likewise.
	* tests/data/test-read-dwarf/test10-pr18818-gcc.so.abi: Likewise.
	* tests/data/test-read-dwarf/test16-pr18904.so.abi: Likewise.
	* tests/data/test-read-dwarf/test22-pr19097-libstdc++.so.6.0.17.so.abi:
	Likewise.
	* tests/data/test-read-dwarf/test9-pr18818-clang.so.abi: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>