ir: Remove an unnecessary comparison
When looking at something else, I noticed that in
method_decl::set_linkage_name, we are unnecessarily testing if the
'this' pointer was equal to the pointer to the member function decl of
a member function looked up from the parent class.

This patch replaces that condition by an assert.  If the 'this'
pointer is different from the pointer to any member function decl
looked up from the parent class, then something is terribly wrong
about the integrity of the IR.

	* src/abg-ir.cc (method_decl::set_linkage_name): Assert that the
	'this' pointer always equals the pointer to any member function
	decl looked up from the parent class.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>