ir: Avoid forgetting potential seemingly duplicated member functions
In some rare cases expressed in DWARF from C++, a member function can
belong to both a declaration-only class (having no data member) in one
translation unit, and to a fully defined class in another translation
unit.  In those cases, the two classes are represented and considered
different.

But then, consider a destructor which mangled name is "D1".  D1 is
represented in both the decl-only class and the fully-defined class.
In the former case, its "this pointer" points to a decl-only class,
while in the later case its "this pointer" points a fully-defined
class.

So, D1 coming from the former case will compare different from the D1
coming from the later case, because of the spurious difference
"decl-only class" versus "fully-defined class".  This is in the
context of a self-comparison.

One way to fix this self-comparison subtle change is to give the
former D1 and the later D1 two different "function ID"s.  Today, the
function ID is just the mangled name, D1.  This patch is defining a
more involved ID which makes the difference between a member function
on a decl-only class and a member function on a fully-defined class.

Note that the only member functions that matter are virtual member
functions because they are the only member functions that are
considered when comparing two classes.

This fixes the self-comparison error found on the binary
gcc-gnat-12.3.1-1.fc37.x86_64/usr/libexec/gcc/x86_64-redhat-linux/12/gnat1
while running the self-comparison command below:

    $ fedabipkgdiff --self-compare -a --from fc37 gcc-gnat

	* src/abg-ir.cc (is_declaration_only_class_or_union_type): Add an
	overload for type_base_sptr.
	(function_decl::get_id): Virtual member functions that are defined
	on decl-only classes have a special ID to avoid confusing them
	with their counterparts defined on defined classes.
	* src/abg-reader.cc (build_function_decl)
	(build_function_decl_if_not_suppressed): Add a new
	add_to_exported_decls parameter to avoid adding the function to
	the set of exported decls to early.  For the function to be
	properly added to the set of exported decls, all its attributes
	must be set, including virtualness.  In some case, those
	attributes are not yet known by the end of the call to
	build_function_decl.  The caller must then set those properties
	and then add the function to the set of exported decls.
	(build_{class,union}_decl): Build the function first, then add its
	missing properties and *then* add it to the set of exported decls.
	(handle_function_decl): Adjust.
	* tests/data/test-read-dwarf/test22-pr19097-libstdc++.so.6.0.17.so.abi: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>