configure,test-diff-pkg.cc: Handle symlinks presence in dist tarball
Since the commit below, there are tests that require symbolic links.

Unfortunately, the tar command used to construct tarballs copies the
target of a symbolic link (and thus doesn't keep the symlink) for the
purpose of greater portability, as some platforms don't support
symbolic links.

This patch uses a tar command that keeps the symlinks in place for the
tests introduced by the commit below to keep working.  On some
platforms however, symlinks are removed from the tarball no matter
what.  In those case, the test that needs symlinks is disabled.

Here is the commit in question:

    bad389f abipkgdiff: Avoid comparing binaries that are outside of the package

	* configure.ac: Use "tar -cf" instead of "tar -chf" to construct
	the tarball.  Detect if the presence of the symlink in the
	distribution and define the WITH_SYMLINKS_KEPT_IN_DIST
	preprocessor macro accordingly.
	* tests/test-diff-pkg.cc (in_out_specs): If symlinks are not
	present in the tarball then the test that requires symlinks is
	deactivated.  Also, fix the
	data/test-diff-pkg/symlink-dir-test1/dir1 test to make it point to
	the target of the symlinks directly, to avoid considering a binary
	twice, in cases where symlink targets are copied.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>