ir,comparison: Represent changed anonymous enums
Now that added/removed non-reachable anonymous enums is supported, we
want to represent changing an anonymous enum.

Strictly speaking, adding or removing an enumerator from an anonymous
enum is represented as deleting of the anonymous enum in the old
state and the addition of an anonymous enum in the new state.

This patch analyses the added/removed anonymous enums and if the old
enum has enumerators contained in the new one, then it assumes we are
looking at anonymous enum change.

	* include/abg-ir.h (is_enumerator_present_in_enum): Declare new
	public function.
	* src/abg-ir.cc (is_enumerator_present_in_enum): Turn this
	static function into a public one.
	* src/abg-comparison.cc
	(corpus_diff::priv::ensure_lookup_tables_populated): Detect that
	an removed/added anonymous enum is actually a changed anonymous
	enum and represent it as such.
	* tests/data/test-abidiff-exit/test-anonymous-enums-change-report-v{0,1}.txt:
	New reference test output files.
	* tests/data/test-abidiff-exit/test-anonymous-enums-change-v{0,1}.c:
	Source code for the some input test binary.
	* tests/data/test-abidiff-exit/test-anonymous-enums-change-v{0,1}.o:
	New test input binaries.
	* tests/data/Makefile.am: Add the new test materials above to
	source distribution.
	* tests/test-abidiff-exit.cc (in_out_specs): Add the new test
	input to the harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>