comparison: Always apply filters on the diff graph
When looking at something else, I noticed that
diff_context::maybe_apply_filters doesn't apply the filters (i.e, the
various passes) on the diff graph if all the categories of changes are
allowed, typically when --harmless is passed to abidiff.

This is wrong because even if all the categories are allowed, the
changes carried by nodes of the diff graph should still be
categorized.  This was an early optimization dating back from the
times where the filtering was slow.

	* src/abg-comparison.cc (diff_context::maybe_apply_filters): Do
	not get out when all the categories of diff changes are allowed.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>