Bug 31045 - Don't try setting translation unit for unique types
Unique types (void, pointer to void and variadic parameter types)
should not have their translation unit set whenever they are being
added to a scope.  This is because they are supposed to be created
independently from any translation unit, even if technically, they are
set to the translation unit they are referenced from, for the first
time.

To handle this, a new function maybe_set_translation_unit is created
to handle the setting of the translation unit for decls added to a
scope by scope_decl::{add,insert}_member_decl.  That new function
asserts that unique types should have their translation unit be set.

	* src/abg-ir.cc (maybe_set_translation_unit): Define new static
	function.
	(scope_decl::{add,insert}_member_decl): Use it.
	* tests/data/test-abidiff-exit/PR31045/zfs-abigail-2.4/libnvpair.{abi,so,suppr}:
	New test input files.
	* tests/data/test-abidiff-exit/PR31045/zfs-abigail-2.4/test-PR31045-report-1.txt:
	New reference test output.
	* tests/data/Makefile.am: Add the new test material above to
	source distribution.
	* tests/test-abidiff-exit.cc (in_out_specs): Add the input above
	to this test harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>
Tested-by: Paolo Pisati <paolo.pisati@canonical.com>