dwarf-reader: Fix detection of C language DIEs
If a DIE comes from C, then for that type, namespace support is
disabled.  In this case, we mistakenly look at the language of the
current translation unit DIE rather than the language of the DIE
itself.  So we mistakenly find that the DIE comes from C when it
actually can come from C++.  And we thus mistakenly disable
namespaces.

Fixed thus.

	* src/abg-dwarf-reader.cc (build_ir_node_from_die): Use
	reader::die_is_in_c rather than looking at the language of the
	current translation unit.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>